requirements.txt
torch
torchvision
efficientnet-pytorch
pandas
matplotlib
tqdm
Pillow
Flask
werkzeug
fpdf